Refactor __init__ into modules and add docstring
