+++ pyintegration/deephaven2/plot/axistransform.py
@@ -1,0 +1,48 @@
+#
+#   Copyright (c) 2016-2022 Deephaven Data Labs and Patent Pending
+#
+""" This module implements the AxisTransform class and related utilites. """
+
+import jpy
+
+from deephaven2 import DHError
+from deephaven2._wrapper_abc import JObjectWrapper
+
+_JAxisTransform = jpy.get_type("io.deephaven.plot.axistransformations.AxisTransform")
+
+_JPlottingConvenience = jpy.get_type("io.deephaven.plot.PlottingConvenience")
+
+AxisTransformNames = list(_JPlottingConvenience.axisTransformNames())
+""" The names of available axis transforms. """
+
+
+class AxisTransform(JObjectWrapper):
+    """ The AxisTransform represents an axis transform that needs to be applied to an axis. """
+    j_object_type = _JAxisTransform
+
+    @property
+    def j_object(self) -> jpy.JType:
+        return self.j_axis_transform
+
+    def __init__(self, j_axis_transform):
+        self.j_axis_transform = j_axis_transform
+
+
+def get_axis_transform_by_name(name: str) -> AxisTransform:
+    """ Returns an AxisTransform object by its name.
+
+    Args:
+        name (str): the predefined AxisTransform name
+
+    Returns:
+        a AxisTransform
+
+    Raises:
+        DHError
+    """
+    try:
+        return AxisTransform(j_axis_transform=_JPlottingConvenience.axisTransform(name))
+    except Exception as e:
+        raise DHError(e, "failed to retrieve the named AxisTransform.") from e
+
+
