+++ pyintegration/deephaven2/plot/selectable_dataset.py
@@ -1,6 +1,8 @@
 #
 #   Copyright (c) 2016-2022 Deephaven Data Labs and Patent Pending
 #
+""" This module defines the SelectableDateSet class and provides two concrete implementations. """
+
 import jpy
 
 from deephaven2._wrapper_abc import JObjectWrapper
@@ -10,6 +12,7 @@ _JTableDefinition = jpy.get_type("io.deephaven.engine.table.TableDefinition")
 
 
 class SelectableDataSet(JObjectWrapper):
+    """ A SelectableDataSet object provides a view of a table on a selectable subset. """
     j_object_type = _JSelectableDataSet
 
     def __init__(self, j_sds):
